Remove miniflare-shared
