+++ fixtures/unsafe-external-plugin/src/plugins/unsafe-service.ts
@@ -7,12 +7,12 @@ import {
 	SERVICE_LOOPBACK,
 	SharedBindings,
 } from "miniflare";
-import { z } from "miniflare-shared";
 // The below imports (prefixed with `worker:`)
 // will be converted by our ESBuild plugin
 // into functions that load the transpiled Workers as JS
 import BINDING_WORKER from "worker:binding.worker";
 import OBJECT_WORKER from "worker:object.worker";
+import { z } from "zod";
 import type { Service, Worker_Binding } from "miniflare";
 
 export const UNSAFE_PLUGIN_NAME = "unsafe-plugin";
