--- packages/miniflare-shared/README.md
@@ -1,3 +1,0 @@
-# Miniflare Shared
-
-This package contains shared code + dependencies for Miniflare.
