--- packages/miniflare-shared/package.json
@@ -1,57 +1,0 @@
-{
-	"name": "miniflare-shared",
-	"version": "4.20250705.0",
-	"description": "Shared code + dependencies for Miniflare",
-	"keywords": [
-		"cloudflare",
-		"workers",
-		"worker",
-		"local",
-		"cloudworker"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/cloudflare/workers-sdk.git",
-		"directory": "packages/miniflare-shared"
-	},
-	"license": "MIT",
-	"main": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"files": [
-		"dist"
-	],
-	"scripts": {
-		"build": "node scripts/build.mjs && pnpm run types:build",
-		"types:build": "tsc --emitDeclarationOnly -d"
-	},
-	"dependencies": {
-		"zod": "3.22.3"
-	},
-	"devDependencies": {
-		"@cloudflare/workers-tsconfig": "workspace:*",
-		"@cloudflare/workers-types": "catalog:default",
-		"@types/node": "catalog:default",
-		"esbuild": "catalog:default",
-		"typescript": "catalog:default"
-	},
-	"peerDependencies": {
-		"esbuild": "catalog:default"
-	},
-	"peerDependenciesMeta": {
-		"esbuild": {
-			"optional": true
-		}
-	},
-	"engines": {
-		"node": ">=18.0.0"
-	},
-	"volta": {
-		"extends": "../../package.json"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"workers-sdk": {
-		"prerelease": true
-	}
-}
