--- packages/miniflare-shared/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./shared/zod.worker";
-export {
-	embedWorkersPlugin,
-	rewriteNodeToInternalPlugin,
-} from "./plugins/esbuild";
