--- packages/miniflare-shared/tsconfig.json
@@ -1,19 +1,0 @@
-{
-	"extends": "@cloudflare/workers-tsconfig/tsconfig.json",
-	"compilerOptions": {
-		"target": "esnext",
-		"lib": ["es2022"],
-		"module": "esnext",
-		"moduleResolution": "bundler",
-		"types": ["@cloudflare/workers-types/experimental", "@types/node"],
-		"noEmit": false,
-		"isolatedModules": true,
-		"allowSyntheticDefaultImports": true,
-		"forceConsistentCasingInFileNames": true,
-		"strict": true,
-		"skipLibCheck": true,
-		"outDir": "dist"
-	},
-	"include": ["src/**/*"],
-	"exclude": ["node_modules", "dist", "**/tests", "**/*.test.ts"]
-}
