--- packages/miniflare-shared/turbo.json
@@ -1,9 +1,0 @@
-{
-	"$schema": "http://turbo.build/schema.json",
-	"extends": ["//"],
-	"tasks": {
-		"build": {
-			"outputs": ["dist/**"]
-		}
-	}
-}
