+++ tools/deployments/__tests__/validate-changesets.test.ts
@@ -38,7 +38,6 @@ describe("findPackageNames()", () => {
 				"edge-preview-authenticated-proxy",
 				"format-errors",
 				"miniflare",
-				"miniflare-shared",
 				"playground-preview-worker",
 				"solarflare-theme",
 				"turbo-r2-archive",
